Fix file structure, create readme
