--- containerized_ml_models/model_api_3/requirements.txt
@@ -1,3 +1,0 @@
-Flask==2.1.3
-numpy==1.23.3
-xgboost==1.5.0
