+++ tasks/01_containerized_ml_models/api/model.pkl
