+++ tasks/01_containerized_ml_models/api/requirements.txt
