+++ tasks/01_containerized_ml_models/model_api_10/model/model.pkl
