+++ tasks/01_containerized_ml_models/model_api_11/model.pkl
