+++ tasks/01_containerized_ml_models/model_api_2/model.pkl
