+++ tasks/01_containerized_ml_models/model_api_3/model_module.py
@@ -1,0 +1,9 @@
+"""
+Using the fact that module is imported only once,
+we get pseudo-singleton behavior by importing model_module
+"""
+import pickle
+
+
+with open("model.pkl", "rb") as file:
+    model = pickle.load(file)
