+++ tasks/01_containerized_ml_models/model_api_3/requirements.txt
@@ -1,0 +1,3 @@
+Flask==2.1.3
+numpy==1.23.3
+xgboost==1.5.0
