+++ tasks/01_containerized_ml_models/model_api_4/model.pkl
