+++ tasks/02_kubernetes_fundamentals/kubernetes_1/README.md
@@ -1,0 +1,36 @@
+# 1. Pomysły i wolne myśli
+
+Podrzucajcie tutaj swoje pomysły, takie które wymagają jakiegoś przypięcia, żeby nie wisiały na discordzie w razie czego.
+
+# 2. TODO
+- null
+- null
+- null
+
+# 3. Podział na zadania
+
+### 3.1 Kuba
+1. null
+2. null
+3. null
+
+
+### 3.2 Mateusz
+1. null
+2. null
+3. null
+
+### 3.3 Adam
+1. null
+2. null
+3. null
+
+# 4. Przydatne linki
+
+[Discord invite](https://discord.gg/bjh2fMFQ)
+
+[Github](https://github.com/Kolo-Naukowe-Data-Science-PW/mlops-project)
+
+[Trello](https://trello.com/c/OQzEVSXb/27-podstawy-kubernetesa-1)
+
+[Kubernetes docs](https://kubernetes.io/pl/docs/concepts/overview/what-is-kubernetes/)
